More additions to the script.js and style.css file
Add universal selector to CSS to make sure layout was right. Add display: flex to make sure the grid layout was correct.

Add button event listener to prompt for an input when the button is pushed to refresh.
Having an issue with the random color generator function. That is the next project.